requirements.txt
tornado
jsonschema
pycurl
python-dateutil